More explictly define groups for Intra_R_Add families
I add explicit H atoms to the groups radadd_intra_CsHNd and radadd_intra_CsHDe, renamed and moved to correct place in tree. I'm not sure if this was the original intent, but the previous version had incorrect sibling parent/child relationship. The group labels are not consistent, and there were no rules with sources to help me make the decision, so I just made my best guess

There is only one rule for each of these families anyhow, so this minor change won't have any effect for now.