Remove deprecated forbidden groups in kinetics families
Many forbidden groups seem to be left over from when we did not have lone pairs on the adj list. Now that we can express the singlet or triplet with 'u' and 'p' variables, these forbidden groups are already implicitly forbidden by the top level group.

The only effect is that they forbid the creation of some products, which does not seem like the original intent.