Delete comments and fix typo in Srad Disproportionation and R_Recombination reactions
Found bug after getting Ea = -50, causing system to
be too stiff and crash.  We found that the alpha and n
kinetics values in Disproportionation were switched
due to a typo. The form of the kinetics should not be
Evans-Polanyi.  We also found that the MRH comments in the
kinetics descriptions are no longer relevant as it
corresponds to a previous version of the rate rule.
The rate rule in the database was constructed by
AG Vandeputte.