TEMP? Making generic "Si" node very high energy.
The idea is that any species which hits this 
top generic node and not a child node, is probably
a crazy species (assuming we entered all the child
nodes correctly from the Wong et al paper) and thus
the formation of said species should be highly endothermic
so it doesn't end up in the core. 
Thus a very bad but high estimate is safer than a bad 
but low estimate!  (or we could raise an exception 
and fix the families that make such species by adding
forbidden structures)