Restructure part of R_Addition_MultipleBond tree for accessibility
Second level restructuring for YJ: changed ordering of children to {HJ, Y_1centerquadrad, Y_1centertrirad, Y_1centerbirad, CJ, OJ, SJ, NJ}, in a more useful order. In particular it is hard to write lack of a lone pair in adjlists, so it is better to put groups with possible lone pairs first.