Added the NitrogenCurran thermo library
This thermo library covers 52 species,
many of which are important in the formation and consumption chemistry
of NOx species in combustion, Calculated at the B3LYP/cc-pVTZ level of theory.

Since RMG does not distinguish between cis\trans conformers, a representative
conformer was chosen based on lower heat of formation.