Added forbidden groups to Intra_H_mig and Intra_ene families to prevent redundancy and unlikely reactions.
Forbid H-shift from the end of a fulvene tail to the far side of the
5-membered ring (and vice versa) in either familiy because TS would be
too strained. Also forbid Intra_ene family from reacting if the site to
which the H is moving is actually a resonantly-stabilized radical site,
in which case the Intra_H_migration family is a more appropriate match.