Updated Deutschmann (Delgado) Nickel reaction library.
I renumbered the reaction indices to be consistent with the Reaction Number in the Delgado
manuscript. I did this to make it easier for me to cross-check things manually. Also, I made
sure that reactions that involve the three vdW species CH4(s), H2O(s), and CO2(s) were
consistently implemented in the associative desorption direction.