Added training reactions to R_Recombination from SOxNOx library
Commented-out training reaction #8 which now has a better rate estimate
(at combustion-relevant temperatures, the williams et al. study is for low T)
Removed training reaction #9 which is incorrect: NO2-2 (O atom has the
rad in this resonance structure) reacts with OH to form HOONO,
not HNO3 as stated in this reaction; this is also not in agreement
with the cited paper by Williams et al.