Corrected fulvene ring() thermo correction, which was no longer consistent with the rest of the database.
This value had not been updated in 10 years, and in that time the group
additivity values for the atoms in fulvene had changed, causing the
overall fulvene enthalpy of formation to be overestimated by ~6 kcal.
This had a noticeable effect on RMG mechanisms, given that fulvene is a
key species in many systems.

The new value is fit to the CBS-QB3 calculations in VinylCPD_H and
Fulvene_H libraries.