Deleted NNH2(S) from thermo_DFT_CCSDF12_BAC
In this library, NNH2 and NNH2(S), which are resonance structures of each
other, were defined with identical thermo. Previously RMG could not
identify this resonance transition. Now that it does, there is no reason
to keep both entries. The correct structure, NNH2, was kept, and the
non-representative structure, NNH2(S), was deleted.