Updated some mid-level tree group definitions in adsorptionPt
Trying to make these more specific than their parent, but more general than their
children. For some, we don't have data. For some, we should perhaps delete a level
(if there is nothing more general than the children). This is not the final word,
but hopefully a step towards it.