Remove incorrect Duplicate merges and flags in NOx2018
I assume some isomorphism checker neglected the multiplicity
and thought CH2 and CH2(S) were isomorphic, and
CH2CHN(S) and CH2CHN, so it merged the kinetics and/or
marked things as duplicates when they shouldn't be.

This commit fixes that, but unfortunately the numbering
is now off and there are multiple reactions with the same index
attribute.