added CoolProp names for solvents
For each solvent, 'name_in_coolprop' is added as a new feature of
solventData.
CoolProp is an external packaged used for fluid properties calculations.
It takes a solvent's name as an input which is saved as
'name_in_coolprop'.
If solvents are not availalbe in CoolProp, its 'name_in_coolprop' is set
to None.