Remove mercaptomethyl
The mercaptomethyl reference enthalpy was parsed incorrectly from the
literature. It actually corresponds to the methylthio radical. The
mercaptomethyl value is not updated because its literature value has
very large uncertainty. Methylthio is not added to the reference
database because it is difficult to converge in C3v symmetry, most
likely due to a Jahn-Teller effect similar to the methoxy radical.