Updated bidentate family reaction parameter estimates in rules.py with literature values
... of similar reaction types.
Note that R42 is HCOO** -> CO2* + H*, which would not be suggested by the Surface_Bidentate_Dissociation reaction recipe. We might want to update this later with a more appropriate estimate.
Additionally, R31 is COOH* + * -> CO2* + H* is not a  Surface_DoubleBond_to_Bidentate reaction, so we might want to update that one later too if we come across better values.