FORM_BOND -> CHANGE_BOND for vdW families
The reverse of FORM_BOND is BREAK_BOND but if you want it to instead
turn it into a van der waals bond, not totally desorb it from the surface.

So in the forward you have to increment the bond from 0, so the reverse
is decrement it to 0 (not break it).

Required some changes to RMG-Py that happen simultaneously with this commit,
in https://github.com/ReactionMechanismGenerator/RMG-Py/pull/2062