Add new solute group files for the solute GAV.
New groups files are added for the solute parameter group additivity
method. The new groups files include `dataCount` for each entry,
which contains the information on the number of data used to fit each
group value.

`abraham.py` and `nonacentered.py` will be no longer used, but they are
not removed just in case someone wants to use the older group files.