added `vdW_radical` to forbidden structures
We restrict vdW adsorbates from having radicals in family templates by restricting that the vdW groups have multiplicity 1.  However, it is diffcult to restrict that vdW products have multiplicity 1 in template products.  Since we do this for each vdw family, we should add `vdW_radical` to globally forbidden structures .If an adsorbate has radical(s), it should chemisorb with surface.