Add more name_in_coolprop for solvents
'name_in_coolprop' is the name of the solvent that is used in a conda
package 'CoolProp'. Temperature-dependence solvation property
calculations are available for the solvents with 'name_in_coolprop'.

These 'name_in_coolprop' are added for more solvents.

Also, previosuly the solvent entry 11 was labelled "dichloroethane" when
it should be "1,1-dichloroethane" (SMILES = CC(Cl)Cl) and it had a wrong
coolprop name. This entry label is fixed to "1,1-dichloroethane" and
its 'name_in_coolprop' is set to None since it is not available in
CoolProp.