Tell CI to use for_rmg branch of RMS
Temorparily we will maintain a branch of ReactionMechanismSimulator.jl
that is specifically for use with RMG.
Thus `for_rmg` is a good name for it.
We used it once before, in August 2023.

This one is/was also called the marcus_development branch.
It includes all the commits necessary to make electrochemistry work
but doesn't include changes for the RMG-RMS interface overhaul.

This commit will be reverted when for_rmg gets merged into RMS main.