Removed H2 + O2 + O2 form primaryH2O2
This reaction is excluded (numbered with an "X") from Konnov 2015. It was added to this library like other selected excluded reactions since RMG may filter it if it is indeed insignificant. A re-examination of the original paper shows that Konnov does not recommend using this reaction at all. Also, it does not make sense in the direction it is written, it could potentially be a legitimate irreversible reaction in the `HO2 + HO2` direction, but other literature sources cited by Konnov 2025 suggest that H2 is not observed from self recombination of HO2.